new test runner, named tests, skeletal reporter
switched from bats to assert.sh (LGPL should be ok to mixin). made some minor modifications to assert.sh for naming tests and hiding the expected result

along with the new test runner there are new test files for each of the puzzle sets right now. a .gitignore to hide the suggested solutions location, `runtests` is working again, and the README has been updated. There is a skeletal reporter dumping out to echo right now.

tests need modification to keep score and report it. need to actually curl to the scoreboard. need to figure out how to report stdout to test players (probably something for davis-ci and not here)